+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Balance
 - Bots will revert to their usual pushing behaviour if a tower is destroyed
+- While bots are forced in defensive mode, they will automatically teleport when below 200 HP. This is to compenstate for their suicidal behaviour
 
 ## 4.2 - Steam Version
 ### Bug Fixing/New Features