+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## 4.3 - Development Version
 ### Bug Fixing/New Features
-- Fixed creep power always activating when pressing load
+- Fixed creep power always activating when pressing load"
+- Hopefully fixed flesh heaps
 
 ### Balance
 - Bots will revert to their usual pushing behaviour if a tower is destroyed